refactor(lambda): rearrange utils locations
